Move miio/airhumidifier_jsqs.py to miio/integrations/humidifier/deerma/
